update: remove index js
